backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
gunicorn>=21.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
python-dateutil>=2.8.0
python-jose[cryptography]>=3.3.0
reportlab>=4.0.0
